+++ internal/server/v1/observations/calculator_test.go
@@ -1,0 +1,53 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package observations
+
+import (
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
+)
+
+func TestVarCalculatorParseFormula(t *testing.T) {
+	strCmpOpts := cmpopts.SortSlices(func(a, b string) bool { return a < b })
+
+	for _, c := range []struct {
+		formula      string
+		wantStatVars []string
+	}{
+		{
+			"Person_Count_Female[ut=NumberUnit;mm=dcAggregate/Census;op=P1Y]/Person_Count[ut=Census]",
+			[]string{"Person_Count_Female", "Person_Count"},
+		},
+		{
+			"Person_Count-Person_Count_Female-Person_Count_Male",
+			[]string{"Person_Count_Female", "Person_Count", "Person_Count_Male"},
+		},
+		{
+			"(Person_Count-Person_Count_Female) / Person_Count_Female",
+			[]string{"Person_Count_Female", "Person_Count"},
+		},
+	} {
+		varCalculator, err := newCalculator(c.formula)
+		if err != nil {
+			t.Errorf("newCalculator(%s) = %s", c.formula, err)
+		}
+		gotStatVars := varCalculator.statVars()
+		if diff := cmp.Diff(gotStatVars, c.wantStatVars, strCmpOpts); diff != "" {
+			t.Errorf("varCalculator.statVars() diff (-want +got):\n%s", diff)
+		}
+	}
+}
