+++ proto/v1/observations.proto
@@ -111,15 +111,15 @@ message DerivedObservationsSeriesRequest {
   // The items of the formula have a form of "StatVar[OptionalFilters]".
   //
   // The following filters are supported:
-  // 1. m: MeasurementMethod
-  // 2. p: ObservationPeriod
-  // 3. u: Unit
-  // 4. s: ScalingFactor
+  // 1. mm: MeasurementMethod
+  // 2. op: ObservationPeriod
+  // 3. ut: Unit
+  // 4. sf: ScalingFactor
   //
   // For example:
   // 1. Person_Count - Person_Count_Female
-  // 2. Person_Count_Female / Person_Count[m=CensusACS5yrSurvey;p=P5Y]
-  // 3. Person_Count - Person_Count_Female - Person_Count_Male[p=P5Y]
+  // 2. Person_Count_Female / Person_Count[mm=CensusACS5yrSurvey;op=P5Y]
+  // 3. Person_Count - Person_Count_Female - Person_Count_Male[op=P5Y]
   // 4. (Person_Count_Male - Person_Count_Female) / Person_Count
   string formula = 2;
 }
